Agregando validacion de ID de MongoDB en el MongoDB Container

+++ Practica Integradora/src/containers/MongoDb.container.js
@@ -1,5 +1,6 @@
 import replace from "../utils/replace.utils.js";
 import asPOJO from "../utils/asPOJO.utils.js";
+import mongoose from "mongoose";
 
 class MongoDbContainer {
   // el constructor recibe el modelo
@@ -17,9 +18,13 @@ class MongoDbContainer {
 
   // uso getAll y filtro por id
   async getOne(id) {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) return undefined;
     const data = await this.model.findOne({ _id: id });
     let product = asPOJO(data);
-    product = replace(product, "_id", "id");
+    if (product) {
+      product = replace(product, "_id", "id");
+    }
     return product;
   }
 
@@ -29,11 +34,15 @@ class MongoDbContainer {
   }
 
   async updateOne(id, itemData) {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) return undefined;
     await this.model.updateOne({ _id: id }, itemData);
     return this.getOne(id);
   }
 
   async deleteOne(id) {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) return undefined;
     return await this.model.deleteOne({ _id: id });
   }
 }
